:ledger: Move var-int to events/
