+++ lib/events/event.js
@@ -2,7 +2,7 @@
 
 var buffer = require('../util/buffer');
 var MalformedError = require('../util/errors').MalformedError;
-var varint = require('../util/var-int');
+var varint = require('../var-int');
 var MetaEvent, SysexEvent, ChannelEvent;
 
 /**
