+++ lib/events/meta-event.js
@@ -5,7 +5,7 @@ var objectAssign = require('object-assign');
 
 var buffer = require('../util/buffer');
 var selectConst = require('../util/select-const');
-var varint = require('../util/var-int');
+var varint = require('../var-int');
 var Event = require('./event');
 
 /**
