+++ lib/events/sysex-event.js
@@ -4,7 +4,7 @@ var util = require('util');
 
 var buffer = require('../util/buffer');
 var selectConst = require('../util/select-const');
-var varint = require('../util/var-int');
+var varint = require('../var-int');
 var Event = require('./event');
 
 /**
